feat(server): add mcp host
