+++ apps/server/api/tests/mod.rs
@@ -1,1 +1,2 @@
+pub mod session_test;
 pub mod stream_test;
