+++ apps/server/api/tests/session_test.rs
@@ -1,0 +1,826 @@
+#[cfg(test)]
+mod tests {
+    use std::{
+        cmp,
+        sync::{
+            Arc,
+            atomic::{AtomicBool, Ordering},
+        },
+        time::Duration,
+    };
+
+    use api::{config, ws::util::audio::pcm_decode};
+
+    use api::ws::{
+        asr::asr_cache::AsrCache,
+        frame::{Frame, FrameResult},
+        llm::llm_cache::LlmCache,
+        session::{
+            Session,
+            listener::{DefaultListener, Listener},
+        },
+        tts::tts_cache::TtsCache,
+        vad::vad_cache::VadCache,
+    };
+    use axum::body::Bytes;
+    use service::chobits::message::{
+        hello::HelloMessage,
+        listen::{ListenMessage, ListenMode, ListenState},
+        tts::TtsState,
+    };
+    use tokio::{sync::Mutex, time::sleep};
+    use tokio_stream::StreamExt;
+    use tracing::info;
+    use tracing_test::traced_test;
+
+    #[tokio::test]
+    #[traced_test]
+    #[ignore]
+    /// hello paramter input and output the hello result
+    /// cargo test --test session_test -- tests::test_chat_flow_hello --ignored --show-output
+    async fn test_chat_flow_hello() {
+        let mut session = create_session().await;
+        session.start().await;
+        let mut output = session.output_frame().await;
+        let join_handle = tokio::spawn(async move {
+            if let Some(Ok(frame_result)) = output.next().await {
+                match frame_result {
+                    FrameResult::HelloResult(_hello_message) => {
+                        return;
+                    }
+                    _ => {
+                        panic!("unexpected frame result");
+                    }
+                }
+            }
+            panic!("receive hello message error");
+        });
+        let hello_frame = Frame::Hello(HelloMessage {
+            ..Default::default()
+        });
+        session.accept_frame(hello_frame.clone()).await;
+        session.stop().await;
+        join_handle.await.unwrap();
+    }
+
+    #[tokio::test]
+    #[traced_test]
+    #[ignore]
+    /// listen voice by manual mode and output the asr text result
+    /// cargo test --features cuda --test session_test -- tests::test_chat_flow_listen_manual --ignored --show-output
+    async fn test_chat_flow_listen_manual() {
+        use std::path::PathBuf;
+
+        let wav_file: PathBuf = [
+            env!("CARGO_MANIFEST_DIR"),
+            "resources",
+            "test",
+            "samples_jfk.wav",
+        ]
+        .iter()
+        .collect();
+        info!("{}", wav_file.display());
+        let (pcm_data, sample_rate) = pcm_decode(wav_file).unwrap();
+        info!(
+            "pcm_data len = {},sample_rate = {}",
+            pcm_data.len(),
+            sample_rate
+        );
+
+        // the follow code is output wav file to test
+        // use wavers::{AudioSample, ConvertSlice, ConvertTo, Samples, read, write};
+        // let fp = "./decode_pcm_data.wav";
+        // let sr: i32 = 16000;
+        // write(fp, &pcm_data, sr, 1);
+
+        const ENCODE_SAMPLE_RATE: u32 = 16000;
+        let mut encoder = opus::Encoder::new(
+            ENCODE_SAMPLE_RATE,
+            opus::Channels::Mono,
+            opus::Application::Audio,
+        )
+        .unwrap();
+
+        // 16000Hz * 1 channel * 60 ms / 1000 = 960
+        const MONO_60MS: usize = ENCODE_SAMPLE_RATE as usize * 60 / 1000;
+        let size = MONO_60MS;
+        info!("size = {}", size);
+        let len = pcm_data.len();
+        let mut count = len / size;
+        if len % size > 0 {
+            count += 1;
+        }
+        info!("count = {}", count);
+        let mut audio: Vec<Vec<u8>> = Vec::new();
+
+        for n in 0..count {
+            let start = n * size;
+            let end = cmp::min((n + 1) * size, len);
+            let packet = encoder
+                .encode_vec_float(&pcm_data[start..end], size)
+                .unwrap();
+            audio.push(packet);
+        }
+
+        let mut session = create_session().await;
+        let session_id = session.id.clone();
+        session.start().await;
+        let mut output = session.output_frame().await;
+        let join_handle = tokio::spawn(async move {
+            while let Some(data) = output.next().await {
+                info!("session id = {}, data = {:?}", session_id, data);
+                match data {
+                    Ok(frame_result) => match frame_result {
+                        FrameResult::HelloResult(_hello_message) => {}
+                        FrameResult::STTResult(_stt_message) => {}
+                        FrameResult::LLMResult(_llm_message) => {}
+                        FrameResult::TTSResult(tts_message) => {
+                            if let Some(state) = tts_message.state
+                                && TtsState::Stop == state
+                            {
+                                return;
+                            }
+                        }
+                        FrameResult::AudioResult(_audio_message) => {}
+                        _ => {
+                            panic!("unexpected frame result");
+                        }
+                    },
+                    Err(_) => {
+                        break;
+                    }
+                }
+            }
+            panic!("receive hello message error");
+        });
+        let hello_frame = Frame::Hello(HelloMessage {
+            ..Default::default()
+        });
+        session.accept_frame(hello_frame.clone()).await;
+        session
+            .accept_frame(Frame::Listen(ListenMessage {
+                state: ListenState::Start,
+                mmod: Some(service::chobits::message::listen::ListenMode::Manual),
+                ..Default::default()
+            }))
+            .await;
+        for n in 0..audio.len() {
+            session
+                .accept_frame(Frame::Voice(Bytes::copy_from_slice(audio.get(n).unwrap())))
+                .await;
+        }
+        session
+            .accept_frame(Frame::Listen(ListenMessage {
+                state: ListenState::Stop,
+                ..Default::default()
+            }))
+            .await;
+        join_handle.await.unwrap();
+        session.stop().await;
+    }
+
+    #[tokio::test]
+    #[traced_test]
+    #[ignore]
+    /// listen voice by auto mode and output the asr text result
+    /// cargo test --features cuda --test session_test -- tests::test_chat_flow_listen_auto --ignored --show-output
+    async fn test_chat_flow_listen_auto() {
+        use std::path::PathBuf;
+
+        let wav_file: PathBuf = [
+            env!("CARGO_MANIFEST_DIR"),
+            "resources",
+            "test",
+            "samples_jfk.wav",
+        ]
+        .iter()
+        .collect();
+        info!("{}", wav_file.display());
+        let (pcm_data, sample_rate) = pcm_decode(wav_file).unwrap();
+        info!(
+            "pcm_data len = {},sample_rate = {}",
+            pcm_data.len(),
+            sample_rate
+        );
+
+        // the follow code is output wav file to test
+        // use wavers::{AudioSample, ConvertSlice, ConvertTo, Samples, read, write};
+        // let fp = "./decode_pcm_data.wav";
+        // let sr: i32 = 16000;
+        // write(fp, &pcm_data, sr, 1);
+
+        const ENCODE_SAMPLE_RATE: u32 = 16000;
+        let mut encoder = opus::Encoder::new(
+            ENCODE_SAMPLE_RATE,
+            opus::Channels::Mono,
+            opus::Application::Audio,
+        )
+        .unwrap();
+
+        // 16000Hz * 1 channel * 60 ms / 1000 = 960
+        const MONO_60MS: usize = ENCODE_SAMPLE_RATE as usize * 60 / 1000;
+        let size = MONO_60MS;
+        info!("size = {}", size);
+        let len = pcm_data.len();
+        let mut count = len / size;
+        if len % size > 0 {
+            count += 1;
+        }
+        info!("count = {}", count);
+        let mut audio: Vec<Vec<u8>> = Vec::new();
+
+        for n in 0..count {
+            let start = n * size;
+            let end = cmp::min((n + 1) * size, len);
+            let packet = encoder
+                .encode_vec_float(&pcm_data[start..end], size)
+                .unwrap();
+            audio.push(packet);
+        }
+
+        let mut session = create_session().await;
+        let session_id = session.id.clone();
+        session.start().await;
+        let mut output = session.output_frame().await;
+        let next_step = Arc::new(AtomicBool::new(false));
+        let next_step_for_sender = next_step.clone();
+        let join_handle = tokio::spawn(async move {
+            let mut count = 0;
+            while let Some(data) = output.next().await {
+                info!("session id = {}, data = {:?}", session_id, data);
+                match data {
+                    Ok(frame_result) => match frame_result {
+                        FrameResult::HelloResult(_hello_message) => {}
+                        FrameResult::STTResult(_stt_message) => {}
+                        FrameResult::LLMResult(_llm_message) => {}
+                        FrameResult::TTSResult(tts_message) => {
+                            let state = tts_message.state;
+                            if let Some(state) = state
+                                && TtsState::Stop == state
+                            {
+                                count += 1;
+                                next_step.store(true, Ordering::Relaxed);
+                                //when next round tts stop after wake tts round
+                                if count >= 2 {
+                                    return;
+                                }
+                            }
+                        }
+                        FrameResult::AudioResult(_audio_message) => {}
+                        _ => {
+                            panic!("unexpected frame result");
+                        }
+                    },
+                    Err(_) => {
+                        break;
+                    }
+                }
+            }
+            panic!("receive hello message error");
+        });
+        session
+            .accept_frame(Frame::Hello(HelloMessage {
+                ..Default::default()
+            }))
+            .await;
+        for n in 0..audio.len() {
+            session
+                .accept_frame(Frame::Voice(Bytes::copy_from_slice(audio.get(n).unwrap())))
+                .await;
+        }
+        session
+            .accept_frame(Frame::Listen(ListenMessage {
+                state: ListenState::Detect,
+                mmod: None,
+                text: Some(String::from("Hello")),
+                ..Default::default()
+            }))
+            .await;
+        session
+            .accept_frame(Frame::Listen(ListenMessage {
+                state: ListenState::Start,
+                mmod: Some(ListenMode::Auto),
+                ..Default::default()
+            }))
+            .await;
+        let mut to_next_step = false;
+        info!("before next step");
+        while !to_next_step {
+            to_next_step = next_step_for_sender.load(Ordering::Relaxed);
+            sleep(Duration::from_millis(500)).await;
+        }
+        info!("after next step");
+        for n in 0..audio.len() {
+            session
+                .accept_frame(Frame::Voice(Bytes::copy_from_slice(audio.get(n).unwrap())))
+                .await;
+        }
+        join_handle.await.unwrap();
+        session.stop().await;
+    }
+
+    #[tokio::test]
+    #[traced_test]
+    #[ignore]
+    /// listen voice by realtime mode and output the asr text result
+    /// cargo test --features cuda --test session_test -- tests::test_chat_flow_listen_realtime --ignored --show-output
+    async fn test_chat_flow_listen_realtime() {
+        use std::path::PathBuf;
+
+        let wav_file: PathBuf = [
+            env!("CARGO_MANIFEST_DIR"),
+            "resources",
+            "test",
+            "samples_jfk.wav",
+        ]
+        .iter()
+        .collect();
+        info!("{}", wav_file.display());
+        let (pcm_data, sample_rate) = pcm_decode(wav_file).unwrap();
+        info!(
+            "pcm_data len = {},sample_rate = {}",
+            pcm_data.len(),
+            sample_rate
+        );
+
+        // the follow code is output wav file to test
+        // use wavers::{AudioSample, ConvertSlice, ConvertTo, Samples, read, write};
+        // let fp = "./decode_pcm_data.wav";
+        // let sr: i32 = 16000;
+        // write(fp, &pcm_data, sr, 1);
+
+        const ENCODE_SAMPLE_RATE: u32 = 16000;
+        let mut encoder = opus::Encoder::new(
+            ENCODE_SAMPLE_RATE,
+            opus::Channels::Mono,
+            opus::Application::Audio,
+        )
+        .unwrap();
+
+        // 16000Hz * 1 channel * 60 ms / 1000 = 960
+        const MONO_60MS: usize = ENCODE_SAMPLE_RATE as usize * 60 / 1000;
+        let size = MONO_60MS;
+        info!("size = {}", size);
+        let len = pcm_data.len();
+        let mut count = len / size;
+        if len % size > 0 {
+            count += 1;
+        }
+        info!("count = {}", count);
+        let mut audio: Vec<Vec<u8>> = Vec::new();
+
+        for n in 0..count {
+            let start = n * size;
+            let end = cmp::min((n + 1) * size, len);
+            let packet = encoder
+                .encode_vec_float(&pcm_data[start..end], size)
+                .unwrap();
+            audio.push(packet);
+        }
+
+        let mut session = create_session().await;
+        let session_id = session.id.clone();
+        session.start().await;
+        let mut output = session.output_frame().await;
+        let join_handle = tokio::spawn(async move {
+            let mut count = 0;
+            while let Some(data) = output.next().await {
+                info!("session id = {}, data = {:?}", session_id, data);
+                match data {
+                    Ok(frame_result) => match frame_result {
+                        FrameResult::HelloResult(_hello_message) => {}
+                        FrameResult::STTResult(_stt_message) => {}
+                        FrameResult::LLMResult(_llm_message) => {}
+                        FrameResult::TTSResult(tts_message) => {
+                            let state = tts_message.state;
+                            if let Some(state) = state
+                                && TtsState::Stop == state
+                            {
+                                count += 1;
+                                //when next round tts stop after wake tts round
+                                if count >= 2 {
+                                    return;
+                                }
+                            }
+                        }
+                        FrameResult::AudioResult(_audio_message) => {}
+                        _ => {
+                            panic!("unexpected frame result");
+                        }
+                    },
+                    Err(_) => {
+                        break;
+                    }
+                }
+            }
+            panic!("receive hello message error");
+        });
+        session
+            .accept_frame(Frame::Hello(HelloMessage {
+                ..Default::default()
+            }))
+            .await;
+        for n in 0..audio.len() {
+            session
+                .accept_frame(Frame::Voice(Bytes::copy_from_slice(audio.get(n).unwrap())))
+                .await;
+        }
+        session
+            .accept_frame(Frame::Listen(ListenMessage {
+                state: ListenState::Detect,
+                mmod: None,
+                text: Some(String::from("Hello")),
+                ..Default::default()
+            }))
+            .await;
+        session
+            .accept_frame(Frame::Listen(ListenMessage {
+                state: ListenState::Start,
+                mmod: Some(ListenMode::RealTime),
+                ..Default::default()
+            }))
+            .await;
+        for n in 0..audio.len() {
+            session
+                .accept_frame(Frame::Voice(Bytes::copy_from_slice(audio.get(n).unwrap())))
+                .await;
+        }
+        join_handle.await.unwrap();
+        session.stop().await;
+    }
+
+    #[tokio::test]
+    #[traced_test]
+    #[ignore]
+    /// get text message and output the asr text result
+    /// cargo test --features cuda --test session_test -- tests::test_chat_flow_handle_text_message --ignored --show-output
+    async fn test_chat_flow_handle_text_message() {
+        let mut session = create_session().await;
+        let session_id = session.id.clone();
+        session.start().await;
+        let mut output = session.output_frame().await;
+        let join_handle = tokio::spawn(async move {
+            while let Some(data) = output.next().await {
+                info!("session id = {}, data = {:?}", session_id, data);
+                match data {
+                    Ok(frame_result) => match frame_result {
+                        FrameResult::HelloResult(_hello_message) => {}
+                        FrameResult::STTResult(_stt_message) => {}
+                        FrameResult::LLMResult(_llm_message) => {}
+                        FrameResult::TTSResult(tts_message) => {
+                            let state = tts_message.state;
+                            if let Some(state) = state
+                                && TtsState::Stop == state
+                            {
+                                return;
+                            }
+                        }
+                        FrameResult::AudioResult(_audio_message) => {}
+                        _ => {
+                            panic!("unexpected frame result");
+                        }
+                    },
+                    Err(_) => {
+                        break;
+                    }
+                }
+            }
+            panic!("receive hello message error");
+        });
+        session
+            .accept_frame(Frame::Hello(HelloMessage {
+                ..Default::default()
+            }))
+            .await;
+        session
+            .accept_frame(Frame::Listen(ListenMessage {
+                state: ListenState::Detect,
+                mmod: Some(service::chobits::message::listen::ListenMode::Manual),
+                text: Some(String::from("Hello")),
+                ..Default::default()
+            }))
+            .await;
+        join_handle.await.unwrap();
+        session.stop().await;
+    }
+
+    #[tokio::test]
+    #[traced_test]
+    #[ignore]
+    /// when a round running and has a break event,the output stream will stop the original output
+    /// cargo test --features cuda --test session_test -- tests::test_chat_flow_break --ignored --show-output
+    async fn test_chat_flow_break() {
+        let mut session = create_session().await;
+        let session_id = session.id.clone();
+        session.start().await;
+        let mut output = session.output_frame().await;
+        let mut count = 0;
+        let join_handle = tokio::spawn(async move {
+            while let Some(data) = output.next().await {
+                info!("session id = {}, data = {:?}", session_id, data);
+                match data {
+                    Ok(frame_result) => match frame_result {
+                        FrameResult::HelloResult(_hello_message) => {}
+                        FrameResult::STTResult(_stt_message) => {}
+                        FrameResult::LLMResult(_llm_message) => {}
+                        FrameResult::TTSResult(tts_message) => {
+                            let state = tts_message.state;
+                            if let Some(state) = state
+                                && TtsState::Stop == state
+                            {
+                                count += 1;
+                                //when next round tts stop after wake tts round
+                                if count >= 2 {
+                                    return;
+                                }
+                            }
+                        }
+                        FrameResult::AudioResult(_audio_message) => {}
+                        _ => {
+                            panic!("unexpected frame result");
+                        }
+                    },
+                    Err(_) => {
+                        break;
+                    }
+                }
+            }
+            panic!("receive hello message error");
+        });
+        session
+            .accept_frame(Frame::Hello(HelloMessage {
+                ..Default::default()
+            }))
+            .await;
+        session
+            .accept_frame(Frame::Listen(ListenMessage {
+                state: ListenState::Detect,
+                mmod: Some(service::chobits::message::listen::ListenMode::Manual),
+                text: Some(String::from("Hello")),
+                ..Default::default()
+            }))
+            .await;
+        session
+            .accept_frame(Frame::Listen(ListenMessage {
+                state: ListenState::Detect,
+                mmod: Some(service::chobits::message::listen::ListenMode::Manual),
+                text: Some(String::from("Hello")),
+                ..Default::default()
+            }))
+            .await;
+        join_handle.await.unwrap();
+        session.stop().await;
+    }
+
+    #[tokio::test]
+    #[traced_test]
+    #[ignore]
+    /// mcp flow and listen voice by realtime mode and output the asr text result
+    ///
+    /// Shell command:
+    /// ``` shell
+    /// cargo test --features cuda --test session_test -- tests::test_mcp_flow_listen_realtime --ignored --show-output
+    /// ```
+    /// 1. [Device -> Server] hello request
+    /// 2. [Server -> Device] hello response
+    /// 3.1.1. [Server -> Device] mcp initialize request
+    /// 3.1.2. [Device -> Server] mcp initialize response
+    /// 3.1.3. [Server -> Device] mcp tools list request
+    /// 3.1.4. [Device -> Server] mcp tools list response
+    /// 3.2.1. [Device -> Server] voice request
+    /// 3.2.2. [Device -> Server] detect wake request
+    /// 4.1.1. [Device -> Server] listen start reqeust
+    /// 4.1.2. [Device -> Server] voice request (loop forever)
+    /// 4.2.0.1. [Server] vad
+    /// 4.2.0.2. [Server] asr
+    /// 4.2.0.3. [Server] llm (user input replace by wake word)
+    /// 4.2.1. [Server -> Device] llm text response (for detect wake word)
+    /// 4.2.2. [Server -> Device] tts response (for detect wake wake word)
+    /// 5.1.0.1. [Server] vad
+    /// 5.1.0.2. [Server] asr
+    /// 5.1.0.3. [Server] llm
+    /// 5.1.0.4. [Server -> Device] mcp call tool(for device call)
+    /// 5.1.0.5. [Device -> Server] mcp call response
+    /// 5.1.0.6. [Server] llm
+    /// 5.1.1. [Server -> Device] llm text response
+    /// 5.1.2. [Server -> Device] tts response
+    async fn test_mcp_flow_listen_realtime() {
+        // TODO: device mcp response
+        let _device_mcp_response: &'static str = r#"
+[
+  {
+    "name": "self.get_device_status",
+    "description": "Provides the real-time information of the device, including the current status of the audio speaker, screen, battery, network, etc.\nUse this tool for: \n1. Answering questions about current condition (e.g. what is the current volume of the audio speaker?)\n2. As the first step to control the device (e.g. turn up / down the volume of the audio speaker, etc.)",
+    "inputSchema": {
+      "properties": {},
+      "type": "object"
+    }
+  },
+  {
+    "name": "self.audio_speaker.set_volume",
+    "description": "Set the volume of the audio speaker. If the current volume is unknown, you must call `self.get_device_status` tool first and then call this tool.",
+    "inputSchema": {
+      "properties": {
+        "volume": {
+          "maximum": 100,
+          "minimum": 0,
+          "type": "integer"
+        }
+      },
+      "required": ["volume"],
+      "type": "object"
+    }
+  },
+  {
+    "name": "self.screen.set_brightness",
+    "description": "Set the brightness of the screen.",
+    "inputSchema": {
+      "properties": {
+        "brightness": {
+          "maximum": 100,
+          "minimum": 0,
+          "type": "integer"
+        }
+      },
+      "required": ["brightness"],
+      "type": "object"
+    }
+  },
+  {
+    "name": "self.screen.set_theme",
+    "description": "Set the theme of the screen. The theme can be `light` or `dark`.",
+    "inputSchema": {
+      "properties": {
+        "theme": {
+          "type": "string"
+        }
+      },
+      "required": ["theme"],
+      "type": "object"
+    }
+  },
+  {
+    "name": "self.camera.take_photo",
+    "description": "Take a photo and explain it. Use this tool after the user asks you to see something.\nArgs:\n  `question`: The question that you want to ask about the photo.\nReturn:\n  A JSON object that provides the photo information.",
+    "inputSchema": {
+      "properties": {
+        "question": {
+          "type": "string"
+        }
+      },
+      "required": ["question"],
+      "type": "object"
+    }
+  }
+]"#;
+
+        // TODO:
+        use std::path::PathBuf;
+
+        let wav_file: PathBuf = [
+            env!("CARGO_MANIFEST_DIR"),
+            "resources",
+            "test",
+            "samples_jfk.wav",
+        ]
+        .iter()
+        .collect();
+        info!("{}", wav_file.display());
+        let (pcm_data, sample_rate) = pcm_decode(wav_file).unwrap();
+        info!(
+            "pcm_data len = {},sample_rate = {}",
+            pcm_data.len(),
+            sample_rate
+        );
+
+        // the follow code is output wav file to test
+        // use wavers::{AudioSample, ConvertSlice, ConvertTo, Samples, read, write};
+        // let fp = "./decode_pcm_data.wav";
+        // let sr: i32 = 16000;
+        // write(fp, &pcm_data, sr, 1);
+
+        const ENCODE_SAMPLE_RATE: u32 = 16000;
+        let mut encoder = opus::Encoder::new(
+            ENCODE_SAMPLE_RATE,
+            opus::Channels::Mono,
+            opus::Application::Audio,
+        )
+        .unwrap();
+
+        // 16000Hz * 1 channel * 60 ms / 1000 = 960
+        const MONO_60MS: usize = ENCODE_SAMPLE_RATE as usize * 60 / 1000;
+        let size = MONO_60MS;
+        info!("size = {}", size);
+        let len = pcm_data.len();
+        let mut count = len / size;
+        if len % size > 0 {
+            count += 1;
+        }
+        info!("count = {}", count);
+        let mut audio: Vec<Vec<u8>> = Vec::new();
+
+        for n in 0..count {
+            let start = n * size;
+            let end = cmp::min((n + 1) * size, len);
+            let packet = encoder
+                .encode_vec_float(&pcm_data[start..end], size)
+                .unwrap();
+            audio.push(packet);
+        }
+
+        let mut session = create_session().await;
+        let session_id = session.id.clone();
+        session.start().await;
+        let mut output = session.output_frame().await;
+        let join_handle = tokio::spawn(async move {
+            let mut count = 0;
+            while let Some(data) = output.next().await {
+                info!("session id = {}, data = {:?}", session_id, data);
+                match data {
+                    Ok(frame_result) => match frame_result {
+                        FrameResult::HelloResult(_hello_message) => {}
+                        FrameResult::STTResult(_stt_message) => {}
+                        FrameResult::LLMResult(_llm_message) => {}
+                        FrameResult::TTSResult(tts_message) => {
+                            let state = tts_message.state;
+                            if let Some(state) = state
+                                && TtsState::Stop == state
+                            {
+                                count += 1;
+                                //when next round tts stop after wake tts round
+                                if count >= 2 {
+                                    return;
+                                }
+                            }
+                        }
+                        FrameResult::AudioResult(_audio_message) => {}
+                        _ => {
+                            panic!("unexpected frame result");
+                        }
+                    },
+                    Err(_) => {
+                        break;
+                    }
+                }
+            }
+            panic!("receive hello message error");
+        });
+        session
+            .accept_frame(Frame::Hello(HelloMessage {
+                ..Default::default()
+            }))
+            .await;
+        for n in 0..audio.len() {
+            session
+                .accept_frame(Frame::Voice(Bytes::copy_from_slice(audio.get(n).unwrap())))
+                .await;
+        }
+        session
+            .accept_frame(Frame::Listen(ListenMessage {
+                state: ListenState::Detect,
+                mmod: None,
+                text: Some(String::from("Hello")),
+                ..Default::default()
+            }))
+            .await;
+        session
+            .accept_frame(Frame::Listen(ListenMessage {
+                state: ListenState::Start,
+                mmod: Some(ListenMode::RealTime),
+                ..Default::default()
+            }))
+            .await;
+        for n in 0..audio.len() {
+            session
+                .accept_frame(Frame::Voice(Bytes::copy_from_slice(audio.get(n).unwrap())))
+                .await;
+        }
+        join_handle.await.unwrap();
+        session.stop().await;
+        todo!();
+    }
+
+    async fn create_session() -> Session<impl Listener> {
+        info!("init vad cahce");
+        VadCache::init().await;
+        info!("init vad cahce successfully");
+        info!("init asr cahce");
+        AsrCache::init().await;
+        info!("init asr cahce successfully");
+        tracing::info!("init llm cahce");
+        LlmCache::init().await;
+        tracing::info!("init llm cahce successfully");
+        tracing::info!("init tts cahce");
+        TtsCache::init().await;
+        tracing::info!("init tts cahce successfully");
+        let vad = VadCache::create_vad();
+        let vad = Arc::new(Mutex::new(vad));
+        let asr = AsrCache::global().instance.clone();
+        let asr = Arc::new(Mutex::new(asr));
+        let close_connection_no_voice_time = config::get().logic().close_connection_no_voice_time();
+        Session::new(
+            Box::new(DefaultListener::new(vad, asr.clone())),
+            Some(close_connection_no_voice_time),
+        )
+    }
+}
